refactor documentation entry point
