+++ apps/documentation/src/components/PostHog/PostHogProvider.tsx
@@ -1,7 +1,7 @@
 import type { JSX, ReactNode } from 'react'
 import { useEffect } from 'react'
 import { PostHogProvider as PostHogLibraryProvider } from 'posthog-js/react'
-import posthog from 'posthog-js'
+import posthogJs from 'posthog-js'
 
 interface PostHogProviderProps {
   children: ReactNode
@@ -12,7 +12,7 @@ export default function PostHogProvider({
 }: PostHogProviderProps): JSX.Element {
   useEffect(() => {
     if (import.meta.env.VITE_ENABLE_POSTHOG === 'true') {
-      posthog?.init(import.meta.env.VITE_PUBLIC_POSTHOG_KEY, {
+      posthogJs.init(import.meta.env.VITE_PUBLIC_POSTHOG_KEY, {
         api_host: import.meta.env.VITE_PUBLIC_POSTHOG_HOST,
         persistence: 'memory', // Cookieless tracking
         loaded: (ph) => {
@@ -23,6 +23,8 @@ export default function PostHogProvider({
   }, [])
 
   return (
-    <PostHogLibraryProvider client={posthog}>{children}</PostHogLibraryProvider>
+    <PostHogLibraryProvider client={posthogJs}>
+      {children}
+    </PostHogLibraryProvider>
   )
 }
