+++ apps/documentation/src/components/PostHog/index.ts
@@ -1,0 +1,4 @@
+import PostHogProvider from './PostHogProvider'
+import PostHogPageView from './PostHogPageView'
+
+export { PostHogProvider, PostHogPageView }
