--- apps/documentation/src/components/PostHogProvider/index.ts
@@ -1,3 +1,0 @@
-import PostHogProvider from './PostHogProvider'
-
-export default PostHogProvider
